Require a valid session to update a users name
While this isn't a security leak, it still allows anyone to
change the name of a Pod owner.

This also makes sure that "If you did not request this
you do not need to take any further action" is correct -
before this fix it should read "If you did not request this
you must still fix your name".